requirements.txt
openai-agents>=0.0.2
python-dotenv>=1.0.0
openai>=1.3.0 